backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
sqlalchemy==2.0.23
marshmallow==3.20.1
python-dotenv==1.0.0
pytest==7.4.3
pyjwt==2.8.0
